Fix Static on negative signals (#1)
Fixed scratchy negatives by passing absolute value into extract_step function

+++ src/types.h
@@ -19,7 +19,7 @@ typedef enum Sign
 typedef struct SignMag
 {
     Sign sign;
-    unsigned int mag;
+    uint16_t mag;
 } SignMag;
 
 #endif
